+++ repository/src/main/java/org/grnet/cat/repositories/registry/CriterionRepository.java
@@ -1,6 +1,5 @@
 package org.grnet.cat.repositories.registry;
 
-import io.quarkus.panache.common.Sort;
 import jakarta.enterprise.context.ApplicationScoped;
 import org.apache.commons.lang3.StringUtils;
 import org.grnet.cat.entities.Page;
@@ -24,9 +23,24 @@ public class CriterionRepository implements Repository<Criterion, String> {
      * @param size   The maximum number of criteria to include in a page.
      * @return A PageQuery of criteria items.
      */
-    public PageQuery<Criterion> fetchCriteriaByPage(int page, int size) {
+    public PageQuery<Criterion> fetchCriteriaByPage(String search, String sort, String order, int page, int size) {
 
-        var panache = find("from Criterion", Sort.by("lastTouch", Sort.Direction.Descending).and("id", Sort.Direction.Ascending)).page(page, size);
+        var joiner = new StringJoiner(" ");
+        joiner.add("from Criterion c");
+
+        var map = new HashMap<String, Object>();
+
+        if (StringUtils.isNotEmpty(search)) {
+            joiner.add("where (c.id like :search")
+                    .add("or c.label like :search")
+                    .add("or c.cri like :search)");
+
+            map.put("search", "%" + search + "%");
+        }
+
+        joiner.add("order by c." + sort + " " + order + ", c.id ASC");
+
+        var panache = find(joiner.toString(), map).page(page, size);
 
         var pageable = new PageQueryImpl<Criterion>();
         pageable.list = panache.list();